feature: centralize GqdCompiledModelGenerator
