+++ projects/pswg_blasters/src/client/java/dev/pswg/datagen/BlasterDataGenerator.java
@@ -2,13 +2,8 @@ package dev.pswg.datagen;
 
 import dev.pswg.Blasters;
 import dev.pswg.BlastersClient;
-import dev.pswg.Galaxies;
-import dev.pswg.data.CodecDataLoader;
-import dev.pswg.data.IdentifierUtil;
 import dev.pswg.item.BlasterItem;
 import dev.pswg.item.HasAttachmentProperty;
-import dev.pswg.rendering.models.GalaxiesModelBakery;
-import dev.pswg.rendering.models.GqbIntermediary;
 import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
 import net.fabricmc.fabric.api.datagen.v1.FabricPackOutput;
@@ -19,9 +14,6 @@ import net.minecraft.client.data.models.ItemModelGenerators;
 import net.minecraft.client.data.models.model.ItemModelUtils;
 import net.minecraft.client.renderer.item.EmptyModel;
 import net.minecraft.core.HolderLookup;
-import net.minecraft.data.CachedOutput;
-import net.minecraft.data.DataProvider;
-import net.minecraft.data.PackOutput;
 import net.minecraft.resources.Identifier;
 import net.minecraft.server.packs.PackType;
 
@@ -46,7 +38,7 @@ public class BlasterDataGenerator implements DataGeneratorEntrypoint
 		pack.addProvider(LangGenerator::new);
 		pack.addProvider(TagGenerator::new);
 		pack.addProvider(ModelGenerator::new);
-		pack.addProvider(GqdCompiledModelGenerator::new);
+		pack.addProvider((fabricPackOutput, completableFuture) -> new dev.pswg.datagen.GqdCompiledModelGenerator(fabricPackOutput, Blasters.MODID));
 	}
 
 	/**
@@ -81,90 +73,6 @@ public class BlasterDataGenerator implements DataGeneratorEntrypoint
 							new EmptyModel.Unbaked()
 					)
 			));
-		}
-	}
-
-	/**
-	 * The GQD compiled model generator. All models should be compiled through
-	 * this generator.
-	 */
-	private static class GqdCompiledModelGenerator implements DataProvider
-	{
-		private final PackOutput.PathProvider resolver;
-
-		public GqdCompiledModelGenerator(FabricPackOutput output)
-		{
-			this.resolver = output.createPathProvider(PackOutput.Target.RESOURCE_PACK, "models");
-		}
-
-		@Override
-		public CompletableFuture<?> run(CachedOutput writer)
-		{
-			var completables = new ArrayList<CompletableFuture<?>>();
-
-			for (var entry : GQB_INTERMEDIARY_LOADER.getDefinitions().entrySet())
-			{
-				if (!entry.getKey().getNamespace().equals(Blasters.MODID))
-					continue;
-
-				completables.add(compile(writer, entry));
-			}
-
-			return CompletableFuture.allOf(completables.toArray(CompletableFuture[]::new));
-		}
-
-		/**
-		 * Compile the given GQB intermediary model
-		 *
-		 * @param writer The writer to add the generated data to
-		 * @param entry  The entry to compile
-		 *
-		 * @return A future that completes when the data is written
-		 */
-		private CompletableFuture<?> compile(CachedOutput writer, Map.Entry<Identifier, GqbIntermediary> entry)
-		{
-			var completables = new ArrayList<CompletableFuture<?>>();
-
-			if (entry.getValue().files().isPresent())
-			{
-				// Split the geometry and model into multiple files
-				for (var fileEntry : entry.getValue().files().get().entrySet())
-				{
-					var nonDatagenId = entry.getKey().withPath(GalaxiesDataProvider.getNonDatagenPath(entry.getKey().getPath(), Optional.of(fileEntry.getKey())));
-					var quadsOutputPath = resolver.file(nonDatagenId, "gqb");
-					var jsonOutputPath = resolver.file(nonDatagenId, "json");
-
-					completables.add(DataProvider.saveStable(writer, entry.getValue().createModelDef(), jsonOutputPath));
-					completables.add(GalaxiesDataProvider.writeToPath(
-							writer,
-							quadsOutputPath,
-							GalaxiesModelBakery.GQuadGeometry.PACKET_CODEC,
-							entry.getValue().createGeometry(Optional.of(new HashSet<>(fileEntry.getValue())))
-					));
-				}
-			}
-			else
-			{
-				var nonDatagenId = entry.getKey().withPath(GalaxiesDataProvider.getNonDatagenPath(entry.getKey().getPath(), Optional.empty()));
-				var quadsOutputPath = resolver.file(nonDatagenId, "gqb");
-				var jsonOutputPath = resolver.file(nonDatagenId, "json");
-
-				completables.add(DataProvider.saveStable(writer, entry.getValue().createModelDef(), jsonOutputPath));
-				completables.add(GalaxiesDataProvider.writeToPath(
-						writer,
-						quadsOutputPath,
-						GalaxiesModelBakery.GQuadGeometry.PACKET_CODEC,
-						entry.getValue().createGeometry(Optional.empty())
-				));
-			}
-
-			return CompletableFuture.allOf(completables.toArray(CompletableFuture[]::new));
-		}
-
-		@Override
-		public String getName()
-		{
-			return "GQD Compiled Models";
 		}
 	}
