+++ projects/pswg_core/src/client/java/dev/pswg/datagen/GalaxiesDataGenerator.java
@@ -69,7 +69,7 @@ public class GalaxiesDataGenerator implements DataGeneratorEntrypoint
 		pack.addProvider(RecipesGenerator::new);
 		pack.addProvider(BlockTagGenerator::new);
 		pack.addProvider(ItemTagGenerator::new);
-		pack.addProvider(GqdCompiledModelGenerator::new);
+		pack.addProvider((fabricPackOutput, completableFuture) -> new dev.pswg.datagen.GqdCompiledModelGenerator(fabricPackOutput, Galaxies.MODID));
 	}
 
 	/**
@@ -438,87 +438,5 @@ public class GalaxiesDataGenerator implements DataGeneratorEntrypoint
 			return "PSWGGalaxiesRecipeProvider";
 		}
 	}
-	/**
-	 * The GQD compiled model generator. All models should be compiled through
-	 * this generator.
-	 */
-	private static class GqdCompiledModelGenerator implements DataProvider
-	{
-		private final PackOutput.PathProvider resolver;
 
-		public GqdCompiledModelGenerator(FabricPackOutput output)
-		{
-			this.resolver = output.createPathProvider(PackOutput.Target.RESOURCE_PACK, "models");
-		}
-
-		@Override
-		public CompletableFuture<?> run(CachedOutput writer)
-		{
-			var completables = new ArrayList<CompletableFuture<?>>();
-
-			for (var entry : GQB_INTERMEDIARY_LOADER.getDefinitions().entrySet())
-			{
-				if (!entry.getKey().getNamespace().equals(Galaxies.MODID))
-					continue;
-
-				completables.add(compile(writer, entry));
-			}
-
-			return CompletableFuture.allOf(completables.toArray(CompletableFuture[]::new));
-		}
-
-		/**
-		 * Compile the given GQB intermediary model
-		 *
-		 * @param writer The writer to add the generated data to
-		 * @param entry  The entry to compile
-		 *
-		 * @return A future that completes when the data is written
-		 */
-		private CompletableFuture<?> compile(CachedOutput writer, Map.Entry<Identifier, GqbIntermediary> entry)
-		{
-			var completables = new ArrayList<CompletableFuture<?>>();
-
-			if (entry.getValue().files().isPresent())
-			{
-				// Split the geometry and model into multiple files
-				for (var fileEntry : entry.getValue().files().get().entrySet())
-				{
-					var nonDatagenId = entry.getKey().withPath(GalaxiesDataProvider.getNonDatagenPath(entry.getKey().getPath(), Optional.of(fileEntry.getKey())));
-					var quadsOutputPath = resolver.file(nonDatagenId, "gqb");
-					var jsonOutputPath = resolver.file(nonDatagenId, "json");
-
-					completables.add(DataProvider.saveStable(writer, entry.getValue().createModelDef(), jsonOutputPath));
-					completables.add(GalaxiesDataProvider.writeToPath(
-							writer,
-							quadsOutputPath,
-							GalaxiesModelBakery.GQuadGeometry.PACKET_CODEC,
-							entry.getValue().createGeometry(Optional.of(new HashSet<>(fileEntry.getValue())))
-					));
-				}
-			}
-			else
-			{
-				var nonDatagenId = entry.getKey().withPath(GalaxiesDataProvider.getNonDatagenPath(entry.getKey().getPath(), Optional.empty()));
-				var quadsOutputPath = resolver.file(nonDatagenId, "gqb");
-				var jsonOutputPath = resolver.file(nonDatagenId, "json");
-
-				completables.add(DataProvider.saveStable(writer, entry.getValue().createModelDef(), jsonOutputPath));
-				completables.add(GalaxiesDataProvider.writeToPath(
-						writer,
-						quadsOutputPath,
-						GalaxiesModelBakery.GQuadGeometry.PACKET_CODEC,
-						entry.getValue().createGeometry(Optional.empty())
-				));
-			}
-
-			return CompletableFuture.allOf(completables.toArray(CompletableFuture[]::new));
-		}
-
-		@Override
-		public String getName()
-		{
-			return "Gadgets GQD Compiled Models";
-		}
-	}
 }
