+++ projects/pswg_gadgets/src/client/java/dev/pswg/datagen/GadgetsDataGenerator.java
@@ -72,7 +72,7 @@ public class GadgetsDataGenerator implements DataGeneratorEntrypoint
 		pack.addProvider(BlockTagGenerator::new);
 		pack.addProvider(ModelGenerator::new);
 		pack.addProvider(RecipesGenerator::new);
-		pack.addProvider(GqdCompiledModelGenerator::new);
+		pack.addProvider((fabricPackOutput, completableFuture) -> new GqdCompiledModelGenerator(fabricPackOutput, Gadgets.MODID));
 	}
 
 	/**
@@ -569,89 +569,6 @@ public class GadgetsDataGenerator implements DataGeneratorEntrypoint
 		public String getName()
 		{
 			return "PSWGGadgetsRecipeProvider";
-		}
-	}
-	/**
-	 * The GQD compiled model generator. All models should be compiled through
-	 * this generator.
-	 */
-	private static class GqdCompiledModelGenerator implements DataProvider
-	{
-		private final PackOutput.PathProvider resolver;
-
-		public GqdCompiledModelGenerator(FabricPackOutput output)
-		{
-			this.resolver = output.createPathProvider(PackOutput.Target.RESOURCE_PACK, "models");
-		}
-
-		@Override
-		public CompletableFuture<?> run(CachedOutput writer)
-		{
-			var completables = new ArrayList<CompletableFuture<?>>();
-
-			for (var entry : GQB_INTERMEDIARY_LOADER.getDefinitions().entrySet())
-			{
-				if (!entry.getKey().getNamespace().equals(Gadgets.MODID))
-					continue;
-
-				completables.add(compile(writer, entry));
-			}
-
-			return CompletableFuture.allOf(completables.toArray(CompletableFuture[]::new));
-		}
-
-		/**
-		 * Compile the given GQB intermediary model
-		 *
-		 * @param writer The writer to add the generated data to
-		 * @param entry  The entry to compile
-		 *
-		 * @return A future that completes when the data is written
-		 */
-		private CompletableFuture<?> compile(CachedOutput writer, Map.Entry<Identifier, GqbIntermediary> entry)
-		{
-			var completables = new ArrayList<CompletableFuture<?>>();
-
-			if (entry.getValue().files().isPresent())
-			{
-				// Split the geometry and model into multiple files
-				for (var fileEntry : entry.getValue().files().get().entrySet())
-				{
-					var nonDatagenId = entry.getKey().withPath(GalaxiesDataProvider.getNonDatagenPath(entry.getKey().getPath(), Optional.of(fileEntry.getKey())));
-					var quadsOutputPath = resolver.file(nonDatagenId, "gqb");
-					var jsonOutputPath = resolver.file(nonDatagenId, "json");
-
-					completables.add(DataProvider.saveStable(writer, entry.getValue().createModelDef(), jsonOutputPath));
-					completables.add(GalaxiesDataProvider.writeToPath(
-							writer,
-							quadsOutputPath,
-							GalaxiesModelBakery.GQuadGeometry.PACKET_CODEC,
-							entry.getValue().createGeometry(Optional.of(new HashSet<>(fileEntry.getValue())))
-					));
-				}
-			}
-			else
-			{
-				var nonDatagenId = entry.getKey().withPath(GalaxiesDataProvider.getNonDatagenPath(entry.getKey().getPath(), Optional.empty()));
-				var quadsOutputPath = resolver.file(nonDatagenId, "gqb");
-				var jsonOutputPath = resolver.file(nonDatagenId, "json");
-
-				completables.add(DataProvider.saveStable(writer, entry.getValue().createModelDef(), jsonOutputPath));
-				completables.add(GalaxiesDataProvider.writeToPath(
-						writer,
-						quadsOutputPath,
-						GalaxiesModelBakery.GQuadGeometry.PACKET_CODEC,
-						entry.getValue().createGeometry(Optional.empty())
-				));
-			}
-
-			return CompletableFuture.allOf(completables.toArray(CompletableFuture[]::new));
-		}
-
-		@Override
-		public String getName()
-		{
-			return "Gadgets GQD Compiled Models";
 		}
 	}
